CAMEL-15948: examples.json instead of .adoc files
This makes use of a change in the camel-quarkus-examples repository to
generate the `examples.json` file instead of empty `.adoc` files to as
a data source for generating the examples table in
`user-guide/examples.adoc`.
It does this by utilizing `@djencks/asciidoctor-jsonpath` Antora
extension, which similarly to the `@djencks/asciidoctor-antora-indexer`
can generate rows for Asciidoc table.