Mark modules using legacy @ConfigRoot as such
The new extension annotation processor coming with 3.14 requires that
you mark the modules using the legacy @ConfigRoot approach (i.e.
without @ConfigMapping) with a specific annotation processor parameter.

This is important for the generation of the config documentation (which
is enabled by default).

This can be merged pre-3.14 update but will result in a warning when
building. With the advantage of having the build with main passing.